Enables -Wall -Werror for building (contributed by Ben Mahler, https://reviews.apache.org/r/6499).
git-svn-id: https://svn.apache.org/repos/asf/incubator/mesos/trunk@1377133 13f79535-47bb-0310-9956-ffa450edef68

+++ src/common/attributes.cpp
@@ -74,7 +74,7 @@ Attributes Attributes::parse(const string& s)
 
   vector<string> tokens = strings::split(s, ";\n");
 
-  for (int i = 0; i < tokens.size(); i++) {
+  for (size_t i = 0; i < tokens.size(); i++) {
     const vector<string>& pairs = strings::split(tokens[i], ":");
     if (pairs.size() != 2) {
       LOG(FATAL) << "Bad value for attributes, missing ':' within " << pairs[0];
